pyproject.toml
[tool.poetry]
name = "greet"
version = "0.1.0"
description = ""
authors = ["jim Cassidy <jim@ion8>"]
readme = "README.md"

[tool.poetry.scripts]
greet = "greet.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.6.1"}
google-cloud-storage = "^2.5.0"
python-dotenv = "^0.21.0"
genson = "^1.2.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"